+++ examples/app/models_autogenerated.py
@@ -51,27 +51,66 @@ class TEmployee(typing.Protocol):
     def __init__(
         self, id: int, name: str, division: str, salary: typing.Optional[float] = None
     ) -> None:
-        """Construct."""
+        """
+        Construct.
+
+        Args:
+            id: Unique identifier for the employee.
+            name: The name of the employee.
+            division: The part of the company the employee works in.
+            salary: The amount of money the employee is paid.
+
+        """
         ...
 
     @classmethod
     def from_dict(
         cls, id: int, name: str, division: str, salary: typing.Optional[float] = None
     ) -> "TEmployee":
-        """Construct from a dictionary (eg. a POST payload)."""
+        """
+        Construct from a dictionary (eg. a POST payload).
+
+        Args:
+            id: Unique identifier for the employee.
+            name: The name of the employee.
+            division: The part of the company the employee works in.
+            salary: The amount of money the employee is paid.
+
+        Returns:
+            Model instance based on the dictionary.
+
+        """
         ...
 
     @classmethod
     def from_str(cls, value: str) -> "TEmployee":
-        """Construct from a JSON string (eg. a POST payload)."""
+        """
+        Construct from a JSON string (eg. a POST payload).
+
+        Returns:
+            Model instance based on the JSON string.
+
+        """
         ...
 
     def to_dict(self) -> EmployeeDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
+        """
+        Convert to a dictionary (eg. to send back for a GET request).
+
+        Returns:
+            Dictionary based on the model instance.
+
+        """
         ...
 
     def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
+        """
+        Convert to a JSON string (eg. to send back for a GET request).
+
+        Returns:
+            JSON string based on the model instance.
+
+        """
         ...
