+++ open_alchemy/models_file/model/sqlalchemy.j2
@@ -12,25 +12,43 @@ class T{{ artifacts.name }}({{ artifacts.parent_cls }}):
     {{ column.name }}: {{ column.type }}{% endfor %}{% endif %}
 
     def __init__(self{{ arg_init_source }}) -> None:
-        """Construct."""
+        """{{ artifacts.init_docstring }}"""
         ...
 
     @classmethod
     def from_dict(cls{{ arg_from_dict_source }}) -> "T{{ artifacts.name }}":
-        """Construct from a dictionary (eg. a POST payload)."""
+        """{{ artifacts.from_dict_docstring }}"""
         ...
 
     @classmethod
     def from_str(cls, value: str) -> "T{{ artifacts.name }}":
-        """Construct from a JSON string (eg. a POST payload)."""
+        """
+        Construct from a JSON string (eg. a POST payload).
+
+        Returns:
+            Model instance based on the JSON string.
+
+        """
         ...
 
     def to_dict(self) -> {{ artifacts.name }}Dict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
+        """
+        Convert to a dictionary (eg. to send back for a GET request).
+
+        Returns:
+            Dictionary based on the model instance.
+
+        """
         ...
 
     def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
+        """
+        Convert to a JSON string (eg. to send back for a GET request).
+
+        Returns:
+            JSON string based on the model instance.
+
+        """
         ...
