+++ tests/open_alchemy/models_file/model/test_source/test_source.py
@@ -18,199 +18,6 @@ _ModelArtifacts = models_file.types.ModelArtifacts
     "artifacts, expected_source",
     [
         (
-            _SQLAlchemyModelArtifacts(
-                name="Model",
-                columns=[],
-                empty=True,
-                arg=_ArgArtifacts(required=[], not_required=[]),
-                parent_cls="Parent",
-            ),
-            '''
-
-class TModel(Parent):
-    """SQLAlchemy model protocol."""
-
-    # SQLAlchemy properties
-    __table__: sqlalchemy.Table
-    __tablename__: str
-    query: orm.Query
-
-    def __init__(self) -> None:
-        """Construct."""
-        ...
-
-    @classmethod
-    def from_dict(cls) -> "TModel":
-        """Construct from a dictionary (eg. a POST payload)."""
-        ...
-
-    @classmethod
-    def from_str(cls, value: str) -> "TModel":
-        """Construct from a JSON string (eg. a POST payload)."""
-        ...
-
-    def to_dict(self) -> ModelDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
-        ...
-
-    def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
-        ...
-
-
-Model: TModel = models.Model  # type: ignore''',
-        ),
-        (
-            _SQLAlchemyModelArtifacts(
-                name="Model",
-                columns=[_ColumnArtifacts(name="column_1", type="type_1")],
-                empty=False,
-                arg=_ArgArtifacts(
-                    required=[
-                        _ColumnArgArtifacts(
-                            name="column_1",
-                            init_type="init_type_1",
-                            from_dict_type="fd_type_1",
-                        )
-                    ],
-                    not_required=[],
-                ),
-                parent_cls="Parent",
-            ),
-            '''
-
-class TModel(Parent):
-    """
-    SQLAlchemy model protocol.
-
-    Attrs:
-        column_1: The column_1 of the Model.
-
-    """
-
-    # SQLAlchemy properties
-    __table__: sqlalchemy.Table
-    __tablename__: str
-    query: orm.Query
-
-    # Model properties
-    column_1: type_1
-
-    def __init__(self, column_1: init_type_1) -> None:
-        """Construct."""
-        ...
-
-    @classmethod
-    def from_dict(cls, column_1: fd_type_1) -> "TModel":
-        """Construct from a dictionary (eg. a POST payload)."""
-        ...
-
-    @classmethod
-    def from_str(cls, value: str) -> "TModel":
-        """Construct from a JSON string (eg. a POST payload)."""
-        ...
-
-    def to_dict(self) -> ModelDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
-        ...
-
-    def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
-        ...
-
-
-Model: TModel = models.Model  # type: ignore''',
-        ),
-        (
-            _SQLAlchemyModelArtifacts(
-                name="Model",
-                columns=[
-                    _ColumnArtifacts(name="column_1", type="type_1"),
-                    _ColumnArtifacts(name="column_2", type="type_2"),
-                ],
-                empty=False,
-                arg=_ArgArtifacts(
-                    required=[
-                        _ColumnArgArtifacts(
-                            name="column_1",
-                            init_type="init_type_1",
-                            from_dict_type="fd_type_1",
-                        ),
-                        _ColumnArgArtifacts(
-                            name="column_2",
-                            init_type="init_type_2",
-                            from_dict_type="fd_type_2",
-                        ),
-                    ],
-                    not_required=[],
-                ),
-                parent_cls="Parent",
-            ),
-            '''
-
-class TModel(Parent):
-    """
-    SQLAlchemy model protocol.
-
-    Attrs:
-        column_1: The column_1 of the Model.
-        column_2: The column_2 of the Model.
-
-    """
-
-    # SQLAlchemy properties
-    __table__: sqlalchemy.Table
-    __tablename__: str
-    query: orm.Query
-
-    # Model properties
-    column_1: type_1
-    column_2: type_2
-
-    def __init__(self, column_1: init_type_1, column_2: init_type_2) -> None:
-        """Construct."""
-        ...
-
-    @classmethod
-    def from_dict(cls, column_1: fd_type_1, column_2: fd_type_2) -> "TModel":
-        """Construct from a dictionary (eg. a POST payload)."""
-        ...
-
-    @classmethod
-    def from_str(cls, value: str) -> "TModel":
-        """Construct from a JSON string (eg. a POST payload)."""
-        ...
-
-    def to_dict(self) -> ModelDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
-        ...
-
-    def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
-        ...
-
-
-Model: TModel = models.Model  # type: ignore''',
-        ),
-    ],
-    ids=["empty", "single column", "multiple column"],
-)
-@pytest.mark.models_file
-def test_sqlalchemy(artifacts, expected_source):
-    """
-    GIVEN model artifacts
-    WHEN sqlalchemy is called with the artifacts
-    THEN the source code for the model class is returned.
-    """
-    source = models_file._model._source.sqlalchemy(artifacts=artifacts)
-
-    assert source == expected_source
-
-
-@pytest.mark.parametrize(
-    "artifacts, expected_source",
-    [
-        (
             _TypedDictArtifacts(
                 required=_TypedDictClassArtifacts(
                     props=[_ColumnArtifacts(name="column_1", type="type_1")],
@@ -597,15 +404,33 @@ class TModel(Parent):
 
     @classmethod
     def from_str(cls, value: str) -> "TModel":
-        """Construct from a JSON string (eg. a POST payload)."""
+        """
+        Construct from a JSON string (eg. a POST payload).
+
+        Returns:
+            Model instance based on the JSON string.
+
+        """
         ...
 
     def to_dict(self) -> ModelDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
+        """
+        Convert to a dictionary (eg. to send back for a GET request).
+
+        Returns:
+            Dictionary based on the model instance.
+
+        """
         ...
 
     def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
+        """
+        Convert to a JSON string (eg. to send back for a GET request).
+
+        Returns:
+            JSON string based on the model instance.
+
+        """
         ...
 
 
@@ -677,15 +502,33 @@ class TModel(Parent):
 
     @classmethod
     def from_str(cls, value: str) -> "TModel":
-        """Construct from a JSON string (eg. a POST payload)."""
+        """
+        Construct from a JSON string (eg. a POST payload).
+
+        Returns:
+            Model instance based on the JSON string.
+
+        """
         ...
 
     def to_dict(self) -> ModelDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
+        """
+        Convert to a dictionary (eg. to send back for a GET request).
+
+        Returns:
+            Dictionary based on the model instance.
+
+        """
         ...
 
     def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
+        """
+        Convert to a JSON string (eg. to send back for a GET request).
+
+        Returns:
+            JSON string based on the model instance.
+
+        """
         ...
 
 
@@ -757,15 +600,33 @@ class TModel(Parent):
 
     @classmethod
     def from_str(cls, value: str) -> "TModel":
-        """Construct from a JSON string (eg. a POST payload)."""
+        """
+        Construct from a JSON string (eg. a POST payload).
+
+        Returns:
+            Model instance based on the JSON string.
+
+        """
         ...
 
     def to_dict(self) -> ModelDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
+        """
+        Convert to a dictionary (eg. to send back for a GET request).
+
+        Returns:
+            Dictionary based on the model instance.
+
+        """
         ...
 
     def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
+        """
+        Convert to a JSON string (eg. to send back for a GET request).
+
+        Returns:
+            JSON string based on the model instance.
+
+        """
         ...
 
 
@@ -857,15 +718,33 @@ class TModel(Parent):
 
     @classmethod
     def from_str(cls, value: str) -> "TModel":
-        """Construct from a JSON string (eg. a POST payload)."""
+        """
+        Construct from a JSON string (eg. a POST payload).
+
+        Returns:
+            Model instance based on the JSON string.
+
+        """
         ...
 
     def to_dict(self) -> ModelDict:
-        """Convert to a dictionary (eg. to send back for a GET request)."""
+        """
+        Convert to a dictionary (eg. to send back for a GET request).
+
+        Returns:
+            Dictionary based on the model instance.
+
+        """
         ...
 
     def to_str(self) -> str:
-        """Convert to a JSON string (eg. to send back for a GET request)."""
+        """
+        Convert to a JSON string (eg. to send back for a GET request).
+
+        Returns:
+            JSON string based on the model instance.
+
+        """
         ...
 
 
@@ -882,8 +761,5 @@ def test_generate(artifacts, expected_source):
     THEN the expected source is returned.
     """
     source = models_file._model._source.generate(artifacts=artifacts)
-
-    print(repr(source))
-    print(repr(expected_source))
 
     assert source == expected_source
